tests: fix multitag tests.
